// client/src/interfaces/UserData.tsx
export interface UserData {
    id: number | null;
    username: string | null;
    email: string | null;
    fName: string | null;
    lName: string | null;
    age: number | null;
    gender: string | null;
    weight: string | null;
    fitnessLevel: string | null;
    fitnessGoals: string | null;
    exercisePreferences: string | null;
  }
  